Allow user to specify axis or button per input in conf file
Users must now specify whether each input is to be an axis 'a' or button 'b' in their conf file by putting 'a' or 'b' at the beginning of each line. This allows note velocity to be mapped to axis value. In order to do this, magic numbers for MIDI statuses 176, 144 and 128 have been removed, and so multi-channel inputs can be read. However, potential negative consequences of this have not yet been worked out and may need fixing in future.